Updated Mixins to bring back the status display while logging in
